feat: [AR] api get order

+++ lib/app/pages/features/detail_page/detail_page_controller.dart
@@ -121,7 +121,7 @@ class DetailPageController extends GetxController {
 
 
 
-  String formatPrice(double price) {
+  String formatPrice(int price) {
     var formattedPrice = NumberFormat.currency(locale: 'id_ID', symbol: 'Rp ').format(price);
     return formattedPrice.replaceAll(",00", "");
   }
